+++ Sphinx/ChangeLog.rst
@@ -2,7 +2,7 @@ ChangeLog
 =========
 * June 7, 2014
 
-  - Added a (first version) cache mechanism to Pygame surfaces used by ``Image``.
+  - Added a cache mechanism to Pygame surfaces used by ``Image``.
   - Added ``Image._url`` attribute.
   - Moved ``_RADIAN_TO_DEGREE``.
   - Print now to ``stderr`` instead ``stdout``.